add component changes
