+++ ubora-cbc-informational/src/Faqs/components/FaqItem/index.css
@@ -1,0 +1,387 @@
+.faq-item{
+    display: flex;
+    flex-direction: column;
+}
+.faq-question{
+    cursor: pointer;
+    font-weight: bold;
+    margin-left: 30%;
+    margin-top: 1.5%;
+    display: flex;
+    gap: 15%;
+    align-items: center;
+    justify-content: space-between;
+    margin-right: 33%;
+}
+
+.faq-question span{
+    background-color: black;
+    color: white; 
+    padding: 2px 7px;
+    border-radius: 50%;
+    margin-left: -165px;  
+
+
+}
+    
+.faq-answer{
+    margin-top: 10px;
+    padding-left: 30px;
+    margin-left: 28.3%;
+    margin-top: 1.5%;
+    width: 36%;
+}
+.line{
+    width: 38.5%;
+    margin-left: 30%;
+    
+}
+
+
+@media screen and (min-width:280px) and (max-width:360px){
+  
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 9.5%;
+        width: 98%;
+        margin-top: 6.5%;
+        display: flex;
+        gap: 15%;
+        justify-content: space-between;
+        margin-right: 6%;
+        font-size: 10px;
+    
+    
+    }
+    
+
+    .faq-question span{
+        background-color: black;
+        color: white; 
+        padding: 2px 7px;
+        border-radius: 50%;
+    
+    
+    }
+   
+    .faq-answer{
+        margin-top: 5px;
+        padding-left: 30px;
+        margin-left: -0.6%;
+        margin-top: 1.5%;
+        width: 78%;
+        font-size: 12px;
+    }
+    .line{
+        width: 99%;
+        margin-left: 9%;
+    }
+        
+}
+
+
+@media screen and (min-width:361px) and (max-width:390px){
+  
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 9%;
+        width: 95.5%;
+        margin-top: 6.5%;
+        display: flex;
+        gap: 15%;
+        justify-content: space-between;
+        margin-right: 10%;
+        font-size: 11px;
+    
+    }
+    
+    .faq-question span{
+        background-color: black;
+        color: white; 
+        padding: 2px 7px;
+        border-radius: 50%;
+    
+    
+    }
+
+    .faq-answer{
+        margin-top: 5px;
+        padding-left: 30px;
+        margin-left: -0.1%;
+        margin-top: 1.5%;
+        width: 78%;
+        font-size: 12px;
+    }
+    .line{
+        width: 97%;
+        margin-left: 9%;
+        
+    }
+
+
+}
+
+
+@media screen and (min-width:400px) and (max-width:450px){
+   
+
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 12%;
+        width: 88%;
+        margin-top: 6.5%;
+        display: flex;
+        gap: 15%;
+        justify-content: space-between;
+        margin-right: 10%;
+        font-size: 11px;
+    
+    }
+  
+    .faq-question span{
+        background-color: black;
+        color: white; 
+        padding: 2px 7px;
+        border-radius: 50%;
+    
+    
+    }
+   
+    .faq-answer{
+        margin-top: 5px;
+        padding-left: 30px;
+        margin-left: 4.8%;
+        margin-top: 1.5%;
+        width: 78%;
+        font-size: 13px;
+    }
+    .line{
+        width: 90%;
+        margin-left: 11%;
+        
+    }
+
+}
+
+
+
+@media screen and (min-width:500px) and (max-width:550px){
+  
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 12%;
+        font-size: 15px;
+        margin-top: 4.5%;
+        display: flex;
+        gap: 15%;
+        justify-content: space-between;
+        margin-right: 1%;
+     
+    }
+    
+    .faq-question span{
+        background-color: black;
+        color: white; 
+        padding: 2px 7px;
+        border-radius: 50%;
+    
+    
+    }
+    
+    .faq-answer{
+        margin-top: 5px;
+        padding-left: 30px;
+        margin-left: 5.7%;
+        margin-top: 1.5%;
+        width: 78%;
+        font-size: 12px;
+    }
+    .line{
+        width: 90%;
+        margin-left: 11%;
+        
+    }
+
+}
+
+
+@media screen and (min-width:720px) and (max-width:800px){
+    .faq-item{
+        margin: 10px 0;
+    }
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 21%;
+        margin-top: 1.5%;
+    }
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 15%;
+        font-size: 20px;
+        margin-top: 3.5%;
+        display: flex;
+        gap: 15%;
+        justify-content: space-between;
+        margin-right: 6%;
+     
+    }
+    
+    .faq-question span{
+        background-color: black;
+        color: white; 
+        padding: 2px 7px;
+        border-radius: 50%;
+    
+    
+    }
+    .faq-answer{
+        margin-top: 5px;
+        padding-left: 30px;
+        margin-left: 10.8%;
+        margin-top: 1.5%;
+        width: 78%;
+    }
+  
+   
+    .line{
+        width: 82%;
+        margin-left: 14%;
+        
+    }
+}
+
+
+@media screen and (min-width:820px) and (max-width:920px){
+    .faq-item{
+        margin: 10px 0;
+    }
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 7%;
+        font-size: 26px;
+        margin-top: 6.5%;
+        display: flex;
+        gap: 15%;
+        justify-content: space-between;
+        margin-right: 0.2%;
+    
+    }
+    
+    .faq-question span{
+        background-color: black;
+        color: white; 
+        padding: 2px 7px;
+        border-radius: 50%;
+    
+    
+    }
+  
+    .faq-answer{
+        margin-top: 8px;
+        padding-left: 30px;
+        margin-left: 5.5%;
+        margin-top: 1.5%;
+        width: 65%;
+        font-size: 20px;
+    }
+ 
+    .line{
+        width: 94%;
+        margin-left: 7%;
+        
+    }
+    
+
+}
+
+
+@media screen and (min-width:1000px) and (max-width:1100px){
+    .faq-item{
+        margin: 10px 0;
+    }
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 13%;
+        font-size: 20px;
+        margin-top: 6.5%;
+        display: flex;
+        gap: 15%;
+        justify-content: space-between;
+        margin-right: 6%;
+        font-size: 28px;
+    
+    }
+    
+    .faq-question span{
+        background-color: black;
+        color: white; 
+        padding: 2px 7px;
+        border-radius: 50%;
+    
+    
+    }
+  
+    .faq-answer{
+        margin-top: 8px;
+        padding-left: 30px;
+        margin-left: 10%;
+        margin-top: 1.5%;
+        width: 65%;
+        font-size: 21px;
+    }
+    .line{
+        width: 85%;
+        margin-left: 12%;
+        
+    }
+
+}
+
+@media screen and (min-width:1200px) and (max-width:1400px){
+    .faq-item{
+        margin: 10px 0;
+    }
+    .faq-question{
+        cursor: pointer;
+        font-weight: bold;
+        margin-left: 21%;
+        font-size: 18px;
+        margin-top: 6.5%;
+        display: flex;
+        gap: 15%;
+        justify-content: space-between;
+        margin-right: 14%;
+        font-size: 24px;
+    
+    }
+    
+    .faq-question span{
+        background-color: black;
+        color: white; 
+        padding: 2px 7px;
+        border-radius: 50%;
+    
+    }
+  
+    .faq-answer{
+        margin-top: 8px;
+        padding-left: 30px;
+        margin-left: 18.6%;
+        margin-top: 1.5%;
+        width: 65%;
+        font-size: 22px;
+    }
+    .line{
+        width: 67%;
+        margin-left: 20%;
+        
+    }
+}
